async_api/requirements.txt
redis==4.4.2
elasticsearch[async]==7.9.1
fastapi==0.93
orjson==3.9.9
pydantic==1.9.0
uvicorn==0.12.2
uvloop==0.17.0
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
